server: move app file to src dir
