+++ packages/server/bin/www.js
@@ -1,5 +1,5 @@
 import http from 'http'
-import app from '../app'
+import app from '../src/app'
 
 require('dotenv').config()
