+++ packages/server/package.json
@@ -4,7 +4,6 @@
   "private": true,
   "author": "Aragon One",
   "license": "(GPL-3.0-or-later OR AGPL-3.0-or-later)",
-  "main": "app.js",
   "engines": {
     "node": ">=9.0.0"
   },
